backend/pyproject.toml
[project]
name = "taxapp"
version = "0.1.0"
description = "Freemium tax preparation engine — Federal + IL + MN with aggressiveness dial and audit risk scoring"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "structlog>=24.4.0",
    "slowapi>=0.1.9",
    # Phase 3: Database + Auth
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "asyncpg>=0.30",
    "passlib[bcrypt]>=1.7",
    "python-jose[cryptography]>=3.3",
    "pyotp>=2.9",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]
ignore = ["RUF002", "RUF003"]  # Allow × (multiplication sign) in comments/docstrings

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
asyncio_default_fixture_loop_scope = "session"
